Move waitForDeviceToReboot to avoid circular dep problem
