+++ src/cmd/device-protection.js
@@ -12,7 +12,6 @@ const FlashCommand = require('./flash');
 const { platformForId } = require('../lib/platform');
 const BinaryCommand = require('./binary');
 const DeviceProtectionHelper = require('../lib/device-protection-helper');
-const { waitForDeviceToReboot } = require('./device-util');
 
 module.exports = class DeviceProtectionCommands extends CLICommandBase {
 	constructor({ ui } = {}) {
@@ -312,17 +311,13 @@ module.exports = class DeviceProtectionCommands extends CLICommandBase {
 			})());
 		} finally {
 			if (putDeviceBackInDfuMode) {
-				await this._waitForDeviceToReboot();
+				await usbUtils.waitForDeviceToReboot();
 				await this.device.enterDfuMode();
 			}
 			if (this.device && this.device.isOpen) {
 				await this.device.close();
 			}
 		}
-	}
-
-	async _waitForDeviceToReboot() {
-		await waitForDeviceToReboot(this.deviceId);
 	}
 
 	/**
