+++ src/cmd/device-protection.test.js
@@ -1,6 +1,7 @@
 const DeviceProtectionCommands = require('./device-protection');
 const { expect, sinon } = require('../../test/setup');
 const deviceProtectionHelper = require('../lib/device-protection-helper');
+const usbUtils = require('./usb-util');
 
 describe('DeviceProtectionCommands', () => {
 	let deviceProtectionCommands;
@@ -313,12 +314,12 @@ describe('DeviceProtectionCommands', () => {
 			const fn = sinon.stub().resolves();
 			deviceProtectionCommands.device.isInDfuMode = true;
 			sinon.stub(deviceProtectionCommands, '_putDeviceInSafeMode').resolves();
-			sinon.stub(deviceProtectionCommands, '_waitForDeviceToReboot').resolves();
+			sinon.stub(usbUtils, 'waitForDeviceToReboot').resolves();
 
 			await deviceProtectionCommands._withDevice({ putDeviceBackInDfuMode: true }, fn);
 
 			expect(deviceProtectionCommands._putDeviceInSafeMode).to.have.been.calledOnce;
-			expect(deviceProtectionCommands._waitForDeviceToReboot).to.have.been.called;
+			expect(usbUtils.waitForDeviceToReboot).to.have.been.called;
 			expect(fn).to.have.been.calledOnce;
 		});
